requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.20.0
accelerate>=0.20.0

# Image processing
Pillow>=9.0.0
matplotlib>=3.5.0

# Jupyter and widgets
ipywidgets>=8.0.0
jupyter>=1.0.0
notebook>=6.4.0

# HTTP and downloads
requests>=2.28.0
urllib3>=1.26.0

# Hugging Face Hub
huggingface-hub>=0.15.0

# Optional GPU acceleration
# Uncomment the following for GPU support:
# torch-audio  # For audio processing if needed
# xformers     # For memory efficient attention

# Development dependencies (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0